Construct Observation function added in observationservice and minor fixes
